move context buttons into list folder
